refactor: extension-manager
