+++ src/app/electron-main/main.ts
@@ -216,14 +216,12 @@ try {
       if (arg.action === 'getApiAccessRules') {
         // 后期加入权限生成，根据moduleID，上层moduleID，应用范围等
         // 或者是像Android, 跳出权限列表让用户自己选择确认放开的权限。
-        const output: string[] = ['getModules', 'getAppModuleList', 'getSlideModuleList', 'hook'];
+        const output: string[] = ['getModules', 'hook'];
         returnValue = output;
       } else if (arg.action === 'getModules') {
         returnValue = moduleManager.getModules(true);
       } else if (arg.action === 'getModule') {
         returnValue = moduleManager.getModule(arg.data.moduleID);
-      } else if (arg.action === 'getAppModuleList') {
-        returnValue = moduleManager.getAppModuleList();
       } else if (arg.action === 'installModule') {
         const data = await moduleManager.installExt(arg.data);
         if (data.code === 0) {
@@ -236,8 +234,6 @@ try {
           // subView.mainView.view.webContents.send('moduleUpdate');
         }
         returnValue = Object.assign(data, { modules: moduleManager.getModules() });
-      } else if (arg.action === 'getSideModuleList') {
-        returnValue = moduleManager.getSideModuleList(subView.appView?.mainModuleID || 'default');
       } else if (arg.action === 'getFeatures') {
         returnValue = moduleManager.getFeatures();
       } else if (arg.action === 'getFeature') {
@@ -252,12 +248,9 @@ try {
         returnValue = configuration.getSettings();
       } else if (arg.action === 'getModuleSettings') {
         returnValue = configuration.getModuleSettings(arg.data.moduleID);
-      } else if (arg.action === 'getSidePosition') {
-        returnValue = subView.appView?.sidePosition;
-        // 获取mock服务地址
       } else if (arg.action === 'getMockUrl') {
+        // 获取mock服务地址
         returnValue = mockServer.getMockUrl();
-        // 重置并初始化mock路由
       } else if (arg.action === 'hook') {
         returnValue = 'hook返回';
       } else if (arg.action === 'getExtensionPagePathByName') {
