--- src/platform/electron-main/browserView/browserView.type.ts
@@ -1,9 +1,0 @@
-import { ViewBounds } from 'eo/shared/common/bounds';
-
-export interface BrowserViewOpts {
-  bounds: ViewBounds;
-  partition?: string;
-  preloadPath?: string;
-  viewPath: string;
-  preload?: string;
-}
