+++ src/platform/node/extension-manager/types/index.ts
@@ -1,3 +1,2 @@
 export * from './handler';
-export * from './manager';
-export * from './loader';
+export * from './module';
