--- src/platform/node/extension-manager/types/loader.ts
@@ -1,18 +1,0 @@
-import {ModuleInfo} from './manager';
-
-/**
- * 模块加载接口.
- */
-export interface ModuleLoaderInterface {
-  loadModules: (modules: Array<ModuleInfo>) => void;
-  loadModule: (module: ModuleInfo) => void;
-}
-
-/**
- * 模块运行环境
- */
-export enum ModuleRuntime {
-  main = 'main',
-  render = 'render',
-  web = 'web'
-}
