--- src/workbench/browser/src/app/shared/types.ts
@@ -1,8 +1,0 @@
-export type FeatureType = {
-  icon: string;
-  label: string;
-  name: string;
-  description: string;
-  key?: string;
-  properties?: object;
-};
